refactor - component folders rearranged to make it more readable
