+++ frontend/src/pages/Homepage_Components/Mainlogo/Mainlogo.tsx
@@ -1,9 +1,8 @@
 import React from 'react'
 import './mainlogo.css'
-import logo from '../../../../images/edited/AEYB_A0_Circle_resized.png'
+import logo from '../../../images/edited/AEYB_A0_Circle_resized.png'
 
 export default function Mainlogo() {
-
     return (
         <img className='logo' src={logo} alt="AEYB logo" />
     )
