+++ frontend/src/pages/Homepage_Components/Menuitem/Menuitem.tsx
@@ -5,13 +5,13 @@ import { faUser } from '@fortawesome/free-solid-svg-icons'
 import { faHouse } from '@fortawesome/free-solid-svg-icons'
 import { faCalendar } from '@fortawesome/free-solid-svg-icons'
 
-export default function Menuitem(prop) {
+export default function Menuitem(prop: { index: number; isClicked: boolean }) {
 
     const icons = [faHouse, faCalendar, faUser]
     const names = ['HOME', 'CALENDAR', 'PROFILE']
 
     return (
-        <div className={prop.isSelected ? 'menuItem selectedItem' : 'menuItem'}>
+        <div className='menuItem'>
             <div className='iconBox'>
                 <FontAwesomeIcon icon={icons[prop.index]} size="2x" className='icon' />
             </div >
