+++ frontend/src/pages/Homepage_Components/sidebar.css
