--- frontend/src/pages/homepage/Home/HomeContent.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function HomeContent() {
-    return (
-        <div>HomeContent</div>
-    )
-}
